Moved execution code to separate file
